uint8_t state -> uint16_t state

+++ src/actions/actions.h
@@ -6,5 +6,5 @@
 
 #include "../telemetry/telemetry.h"
 
-void handleActions(const Telemetry &telemetryData, uint8_t &buttonsState,
-                   uint8_t &ledState);
+void handleActions(const Telemetry &telemetryData, uint16_t &buttonsState,
+                   uint16_t &ledState);
